mi0.5.0 - Better SQL dependency management
Now `get_HH` function is fully independent of implementations. Corner
cases are not contemplated except for a `None` annotation as a
possible returned value from its child functions.

"Drivers" added as `sqlite.py` and `sqlsrv.py`, and are picked based
on the contents of `cogs.json`.

PyODBC was added as a new default dependency. Although, this can't be
tested in Linux. Worse, this can only be tested in some specific
windows machines.

Given the above text, `sqlsrv.py` must be finished outside the
testing environment. So the project is excepted to explode everywhere
else if not using `sqlite.py`.

     modified:   src/app/pedales/queries/sql.py
     new file:   src/app/pedales/queries/sqlite.py
     new file:   src/app/pedales/queries/sqlsrv.py
     modified:   src/app/settings/cogs.json
     modified:   src/app/settings/req.txt